让识别异常不再中断正式比赛流程
保留可信视觉结果，并对缺失豆子和数字箱进行互斥补全；比赛页面实时展示阶段与识别结果，增加故障历史和受约束的瞬态重试。结合代码审查收紧运动故障恢复边界，并调整K230、步进轴和底盘导航超时。

Constraint: 比赛流程必须在视觉结果不完整时继续，但不得绕过急停、限位、坐标和载荷安全检查
Rejected: 对所有导航和抓放故障自动重做 | 运动后坐标或载荷状态可能已经不可信
Confidence: high
Scope-risk: moderate
Directive: 修改推定映射或恢复策略时，必须验证三种豆子目标箱互斥及运动故障后的坐标可信性
Tested: Debug与Release构建通过；git diff --check通过
Not-tested: 完整场地三轮抓放与K230断线降级实机测试

+++ Core/Inc/font.h
@@ -22,7 +22,7 @@ typedef struct Font {
   uint8_t h;              // 字高度
   uint8_t w;              // 字宽度
   const uint8_t *chars;   // 字库 字库前4字节存储utf8编码 剩余字节存储字模数据
-  uint8_t len;            // 字库长度 超过256则请改为uint16_t
+  uint16_t len;           // 字库长度，支持超过256个中文字模
   const ASCIIFont *ascii; // 缺省ASCII字体 当字库中没有对应字符且需要显示ASCII字符时使用
 } Font;
