+++ Core/Src/oled.c
@@ -812,7 +812,7 @@ void OLED_PrintString(uint8_t x, uint8_t y, char *str, const Font *font, OLED_Co
       break; // 有问题的UTF-8编码
 
     // 寻找字符  TODO 优化查找算法, 二分查找或者hash
-    for (uint8_t j = 0; j < font->len; j++)
+    for (uint16_t j = 0; j < font->len; j++)
     {
       head = (uint8_t *)(font->chars) + (j * oneLen);
       if (memcmp(str + i, head, utf8Len) == 0)
